R code for Analyzing the Prestige Dataset
The goal of the project is to find relationship between occupation,
income, education and sex. This project involves conducting data
analysis on the Prestige dataset from the website
https://vincentarelbundock.github.io/Rdatasets/datasets.html, which is
a dataset of the observations of occupations in the year 1971 of the
Canadian population. The aspects explored in this study are how the
prestige of occupation, average income, percentage of women in
occupation, and type of occupation relate to each other using an
exploratory analysis. Along with this, I used bivariate and
multivariate regression and two statistical analysis techniques on the
data in R which led us to the conclusion that there is a relationship
between education and income, prestige and income, and women and
income. This results shows show how the sex of a person used to play
and still plays a vital role in determining a prestige of a job or
income and how education affects income and prestige of an occupation.